+++ tests/Integration/McpTransportTest.php
@@ -120,5 +120,6 @@ final class McpTransportTest extends TestCase
         $setting->set('value', $value);
         $setting->save();
         self::$modx->config[$key] = $value;
+        self::$modx->getCacheManager()->refresh(['system_settings' => []]);
     }
 }